fix tests
These tests were missing the `error` in the response; the intended behavior should have actually been an "empty cursor" but instead we received a functionally-equivelant `null` and an error due to the bug.

This generally refactors these tests to use an approach more like a "snapshot" than individual checks against the response. This can be a pain when updating implementation details (such as internal cursor formats) but is much more likely to catch cases that would otherwise be missed.