Memoize Authorization.access
Because Authorization.access is called at least once for every entity returned in a request, it is a very hot path. Computing scope intersections is a reasonably expensive operation, and the async nature of the backing entities (grant, user, roles) makes this method relatively slow even when those are memoized and db calls can be avoided.

An executor is already the context used for memoization of database calls, and this expands that to include memoization of calls to `Authorization.access`.